File input output some minor change + checked

+++ 9-File_Iinp_Outp-and-exceptions/File_input-output/File_input-output/Attack.cs
@@ -32,8 +32,8 @@ public class Attack
         get { return kind; }
     }
 
-    private readonly int power;
-    public int Power
+    private readonly double power;
+    public double Power
     {
         get { return power; }
     }
